Load all YAML fragments from all files in a directory

These function extracts all YAML fragments from all files in a directory returning a list of character vectors containing the extracted fragments.

load_yaml_dir(path, recursive = TRUE, fileRegexes = c("^[^\\.]+.*$"),
  select = ".*", delimiterRegEx = "^---$",
  ignoreOddDelimiters = FALSE, encoding = "UTF-8", silent = TRUE)

Arguments

path
The path containing the files.

recursive
Whether to also process subdirectories (TRUE) or not (FALSE).

fileRegexes
A vector of regular expressions to match the files against: only files matching one or more regular expressions in this vector are processed. The default regex (^[^\.]+.*$) matches all files except those that start with a period (.).

delimiterRegEx
The regular expression used to locate YAML fragments.

ignoreOddDelimiters
Whether to throw an error (FALSE) or delete the last delimiter (TRUE) if an odd number of delimiters is encountered.

encoding
The encoding to use when calling readLines(). Set to NULL to let readLines() guess.

silent
Whether to be silent (TRUE) or informative (FALSE).

Value

A list of lists of objects.

Details

These function extracts all YAML fragments from all files in a directory and then calls yaml::yaml.load() to parse them. It then returns a list where each element is a list with the parsed fragments in a file.

Examples

yum::load_yaml_fragments(text=c("---", "First: YAML fragment", "---",
                                "Outside of YAML",
                                "---", "Second: YAML fragment", "---",
                                "Also outside of YAML"));

#> [[1]]
#> [[1]]$First
#> [1] "YAML fragment"
#> 
#> 
#> [[2]]
#> [[2]]$Second
#> [1] "YAML fragment"
#> 
#> 
#> attr(,"class")
#> [1] "yumFromFile" "list"