Convert the objects loaded from YAML fragments into a tree

If the data.tree::data.tree package is installed, this function can be used to convert a list of objects, as loaded from extracted YAML fragments, into a data.tree::Node().

build_tree(x, idName = "id", parentIdName = "parentId",
  childrenName = "children", autofill = c(label = "id"),
  silent = TRUE)

Arguments

x
Either a list of YAML fragments loaded from a file with load_yaml_fragments(), or a list of such lists loaded from all files in a directory with load_yaml_dir().

idName
The name of the field containing each elements' identifier, used to build the data tree when there are references to a parent from a child element.

parentIdName
The name of the field containing references to an element's parent element (i.e. the field containing the identifier of the corresponding parent element).

autofill
A named vector where the names represent fields to fill with the values of the fields specified in the vector values. Note that autofill replacements are only applied if the fields to be autofilled (i.e. the names of the vector specified in autofill) do not already have a value.

silent
Whether to provide (FALSE) or suppress (TRUE) more detailed progress updates.

childName
The name of the field containing an element's children, either as a list of elements, or using the 'shorthand' notation, in which case a vector is supplied with the identifiers of the children.

Value

a data.tree::Node() object.

Examples

Add example here!

#> Error: <text>:1:5: unexpected symbol
#> 1: Add example
#>         ^